Basic error system, along with minor things: removed boost dependance, size_t instead of int for loops etc.

+++ src/construct_debug.cpp
@@ -37,7 +37,7 @@ std::string token_to_string(con_token token) {
       break;
     case FUNCTION:
       tokstring += ", function: " + token.tok_function->name + ", arguments: ";
-      for(int i = 0; i < token.tok_function->arguments.size(); i++) {
+      for(size_t i = 0; i < token.tok_function->arguments.size(); i++) {
         if(i != 0) {
           tokstring += ", ";
         }
@@ -61,7 +61,7 @@ std::string token_to_string(con_token token) {
   }
   if(token.tokens.size() > 0) {
     tokstring += ", tokens: {\n";
-    for(int i = 0; i < token.tokens.size(); i++) {
+    for(size_t i = 0; i < token.tokens.size(); i++) {
       tokstring += token_to_string(*token.tokens[i]) + "\n";
     }
     tokstring += "}";
